connecting the slider (gradientwidget) and the elementproperty class with double insteat of int. so other classes than gradientwidget don't have to know about the Multiplikator.
svn path=/trunk/KDE/kdeedu/kalzium/; revision=1165322

+++ src/gradientwidget_impl.h
@@ -52,7 +52,7 @@ public:
     /**
      * Is emitted when the value of the gradient or the spinbox is changed.
      */
-    void gradientValueChanged( int );
+    void gradientValueChanged( double );
 
 public slots:
     /**
@@ -70,7 +70,7 @@ private slots:
     /**
      * is used to display custom text in the widget accrding tu the current value of the gradient
      */
-    void setNewValue(int newValue);
+    void setNewValue(double newValue);
 
 private:
